test: Move MockFactory into its own package
