+++ org.osgi.test.cases.component/src/org/osgi/test/cases/component/junit/DeclarativeServicesControl.java
@@ -82,8 +82,8 @@ import org.osgi.test.cases.component.service.TestSet;
 import org.osgi.test.cases.component.tb13.ModifyRegistrator;
 import org.osgi.test.cases.component.tb13a.ModifyRegistrator2;
 import org.osgi.test.cases.component.tb6.ActDeactComponent;
-import org.osgi.test.support.MockFactory;
 import org.osgi.test.support.compatibility.DefaultTestBundleControl;
+import org.osgi.test.support.mock.MockFactory;
 import org.osgi.test.support.sleep.Sleep;
 import org.osgi.util.promise.Promise;
 import org.osgi.util.tracker.ServiceTracker;
