+++ org.osgi.test.support/src/org/osgi/test/support/mock/MockFactory.java
@@ -16,7 +16,7 @@
  * SPDX-License-Identifier: Apache-2.0 
  *******************************************************************************/
 
-package org.osgi.test.support;
+package org.osgi.test.support.mock;
 
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
